move and rename typings
