+++ bitmovin-analytics-datasource/src/module.ts
@@ -2,7 +2,7 @@ import { DataSourcePlugin } from '@grafana/data';
 import { DataSource } from './datasource';
 import { ConfigEditor } from './components/ConfigEditor';
 import { QueryEditor } from './components/QueryEditor';
-import { BitmovinAnalyticsDataQuery, BitmovinDataSourceOptions } from './types';
+import { BitmovinAnalyticsDataQuery, BitmovinDataSourceOptions } from './types/grafanaTypes';
 
 export const plugin = new DataSourcePlugin<DataSource, BitmovinAnalyticsDataQuery, BitmovinDataSourceOptions>(
   DataSource
